unify cargo stylus subcommands
