--- cgen/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "cargo-stylus-cgen"
-keywords = ["arbitrum", "ethereum", "stylus", "alloy", "gdb"]
-description = "CLI tool generating C bindings for Arbitrum Stylus ABIs"
-
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[dependencies]
-alloy-json-abi.workspace = true
-clap.workspace = true
-eyre.workspace = true
-serde_json.workspace = true
-tiny-keccak.workspace = true
