+++ main/src/project.rs
@@ -9,7 +9,7 @@ use crate::{
     macros::*,
 };
 use brotli2::read::BrotliEncoder;
-use cargo_stylus_util::{color::Color, sys};
+use crate::util::{color::Color, sys};
 use eyre::{bail, eyre, Result, WrapErr};
 use glob::glob;
 use std::{
