+++ main/src/util/text.rs
