+++ main/src/wallet.rs
@@ -1,12 +1,12 @@
 // Copyright 2023-2024, Offchain Labs, Inc.
 // For licensing, see https://github.com/OffchainLabs/cargo-stylus/blob/main/licenses/COPYRIGHT.md
 
+use crate::util::text::{self, decode0x};
 use crate::AuthOpts;
 use alloy_primitives::FixedBytes;
 use alloy_provider::network::EthereumWallet;
 use alloy_signer::Signer;
 use alloy_signer_local::PrivateKeySigner;
-use cargo_stylus_util::text::{self, decode0x};
 use ethers::signers::LocalWallet;
 use eyre::{eyre, Context, Result};
 use std::fs;
