--- util/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "cargo-stylus-util"
-keywords = ["arbitrum", "ethereum", "stylus", "alloy", "cargo"]
-description = "utilities for cargo stylus and its dependencies"
-
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-version.workspace = true
-repository.workspace = true
-
-[dependencies]
-hex.workspace = true
-ethers.workspace = true
-eyre.workspace = true
-rustc-host.workspace = true
